mpv: select 720p for streaming
720p works very well with FSRCNNX, and the lower resolution has the
advantage of significantly reducing the time to initially load a video
as well as the time it takes to seek around within a video.

This doesn't affect the videos I download, which probably account for
most of the videos that I watch, so the loss in quality shouldn't be too
bad, and AI upscaling can help fill the gap.