auto-save-state: workaround for reloading files w/o seeking to 0
This adds a command that skips deleting the state for the currently
playing file. end-file's "stop" reason is vague, it could refer to stop,
playlist-next, loadfile, and more. So instead of determining which
"stops" should delete state, push the problem to the user by giving the
ability to skip the next delete.