nlmeans.glsl: retire bilateral
I realized the patchwise component of bilateral was incorrect, since it
was getting the average difference without accounting for negatives. So
two unique patches with the same intensity could be viewed as identical
despite that not being the case.

Yikes. Realizing this, the correct implementation would be to use
pow/abs, which virtually yields NLM. So might as well just use NLM.