nlmeans.glsl: fix chroma redundant component & better optimize
Chroma was using a redundant component due to silly error in the macro.

Now that is fixed & the two components are packed into a uint for
storage in the WD=2 array for even better performance (it's only a small
increase, for me it was ~0.1ms faster on yuv444p 1080p video).