shader_test: merge calls to ffmpeg for noise realizations
This makes tests start much faster. It has a side effect of producing
slightly higher SSIMs for some reason. I'm not sure why. I checked to
make sure the output looks OK and all of the realizations are there, but
something must be different?

Either way, this doesn't appear to affect the test results when
comparing them against each other, but they shouldn't be compared
against test results prior to this commit.