shader_test: batch tests during binary search
Should be significantly faster when the shader isn't demanding, since
most of the time is spent spinning ffmpeg/libplacebo up and down.

The only downsides are that SSIMs seem to be very slightly different
(I'm really not sure why, but the difference doesn't appear to be big
enough to be an issue) and the maximizer can go 1 iteration over its
user specified limit (might be annoying when the shader is very slow).

Technically the first two iterations can be batched too, I will
implement that after refactoring the maximizer.