nlmeans.glsl: rework sharpening
The unsharp mask is computed differently now. Instead of being a
comparison between the input image and the NLM blur (in a perfect world
that would just be noise), now it is a comparison between the NLM blur
and the spatial kernel's blur.

This makes the unsharp mask itself noise resistant, rather than relying
on the weight map to control sharpness. This looks a lot better, mostly
since noise is no longer amplified in edge areas.

As a side effect, the ASP, ASC, and ASK options are obsolete and
therefore removed. I also removed ASW since it wasn't very useful.

I am open to adding back something like ASP in the future, but using the
weight map directly works pretty well so I'm content with leaving it
that way for now.

I also added a visualization mode for viewing the unsharp mask and
another one for viewing the effect of EP.