hdeband: implement textureGather
Implementation is kinda broken (probably just sampling in the wrong
place somewhere) but it's not any faster than the non-gather
implementation (maybe because it only uses half of the samples) So this
is probably never going to hit the master branch.

Maybe it's possible to implement this with textureGatherOffsets? I can't
see any way to do that without ridiculous code generation though. I
tried a constant array of arrays but the compiler doesn't like that.