hdeband.glsl: use averages of runs
The idea behind this is to better identify runs by comparing them to an
average instead of just the previous value, in order to better account
for small variations.

But this appears to actually require a higher TOLERANCE setting in
order to achieve the same amount of blur. I'm guessing that's because
this is running into issues with floating point precision, but I could
have just made a mistake in my implementation.