nlmeans.glsl: bring back PS=4:P=3 textureGather
Originally I removed this because I thought it wasn't worth maintaining
considering it is worse quality than PS=3 while having about the same
performance. After thinking about it more, I think the difference in
performance may be more noticeable on lower end systems, and since the
implementation is simple it should be easy enough to maintain.