nlmeans_cfg: sort into LQ/HQ/HQX directories
I plan to add more of these types of shaders, so this will help with
clutter.

I also removed the "heavy" variants of these shaders, which I had been
planning to do for a while. This is because they don't appear to offer a
realistic benefit over the "medium" variants.