nlmeans.glsl: implement plus X shape
Should be a pretty good shape. It's dense in the center where samples
tend to be more important, and sparsity increases away from the center
where samples tend to be less important.

I ran a quick test, maximizing S and SS, and I did measure a slight
improvement compared to the default diamond shape, so at the very least
it works.