Makefile: tweak nlmeans_2x
Finally it is performing above baseline. There's not really a secret as
to why, it's the sum of many small optimizations. Also I am testing it
for only upscaling now, rather than upscale+denoise (which it was above
baseline for, but only because it did a good job on the denoise part).
Upscale+denoise can be added in the future.

Also this doesn't do chroma yet.