shader_test: new optimizer, WIP
Right now it works by pretty much just adding a bunch of random offsets
to the estimate (starting with the starting parameters) and then
replacing the estimate with the best sample.

This is probably already superior to the old optimizer in some ways, but
it needs some type of decay in order fully replace it.