nlmeans.glsl: always skip POI in research loop (research donuts)
Previously this was only done if no RF was being used. I believe this
was because of the way RF was (incorrectly) implemented before (I can't
remember the order of events and it's probably not worth digging around
in git).

Either way, a redundant sample/comparison was being done, so this commit
brings free speed.

Also I demoted the LGC shaders to the experimental directory, they
haven't been tested in a long time and I'm not convinced that LGC is the
way to go for upscaling chroma, at least not the way it exists now.