nlmeans.glsl: parameterize weight discard
Now weight discard has the capability to decrease weights instead of
just zeroing them out. Not sure how useful this may or may not be in
practice, I haven't tested that yet, I've only ensured that this commit
doesn't introduce any regressions.

So with the default settings this should behave virtually identical to
before with only two exceptions:

1. I fixed a (very small) bug in WD=1, there was an off-by-one. This had
   no measurable effect in practice though.

2. The final state of avg_weight decreases instead of increasing now.
   The math is much simpler this way. This currently has no effect in
   practice since this variable in only really used within WD (excluding
   weight map visualization).