nlmeans.glsl: use a separate kernel for sharpening
Previously the spatial kernel (SK) was being reused along with the
spatial sigma (SS). Now there are separate parameters that are specific
to sharpening (ASK & ASS).

This makes AS practical for upscaling now, since an appropriate amount
of blur can be used for the unsharp mask (since in that case the spatial
kernel's job is to faithfully reconstruct the image rather than to blur
it). Although I have only been able to find a mild improvement through
this.