nlmeans.glsl: allow for sharpening noise
The idea being that this could be useful in faithfully upscaling noisy
content. Although this doesn't perform any better in my benchmarks since
the dataset is not noisy, and failing to reconstruct noise is probably
not going to be seen as much of a failure outside of benchmarks.

So realistically this option isn't very useful.