shaders/Makefile: fix some options not being set properly
C=0 was being set for the luma plane of a few shaders when it should
have been C=1. The difference is imperceptible, but it's fixed and I
added a script (ensure_opt_correctness) to check if this happens again
in the future. It's only a workaround, ideally the options should be
set/interpreted in a smarter way so that stuff like this doesn't happen.