nlmeans.glsl: implement new visualization mode, celled weight map
This visualizes the individual weights of a subset of pixels. This
should hopefully help to understand the shaders for future improvements
and debugging.

Like many of the other visualizations, this is luma-only for now.
Ideally this should be used with --scale=nearest and a lot of zoom.

Also the existing weight maps were renamed to be average weight maps in
order to better describe them and distinguish them from this new weight
map.