shader_cfg: big oops, fix chroma inject opts overriding luma
I believe this has been an issue since f33d32d. This affects the
injected luma plane of the shaders built with shader_cfg whenever a
chroma plane was also specified (e.g., most of the shaders but not the
tests, since only one plane is tested at a time).

Related to #11