shaders/Makefile: fix regression in sharpening quality on default
The regression was introduced in c84ce25. The size of the research zone
was reduced because it was hardly being used in denoising since the
guide was doing most of the work, however all of the sharpening was
still being handled by the last pass, now with a small research size.

This commit fixes this by doing the sharpening on both passes.