hdeband: fix potential issue & remove sparsity compensation
It was possible for the previous weight to get accumulated multiple
times, although this doesn't appear to have had any significant effect
in practice, at least not with the default settings.

The sparsity compensation code was removed for simplicity sake. It
wasn't implemented very well and sparsity isn't very useful. Sparsity is
still implemented, it just works in a slightly simpler fashion now.