auto-save-state: save on file-loaded if playlist-pos > 0
Without this there would be a gap (up to 60 seconds) between autosaves
after loading the next entry of the playlist, which would result in the
playlist's watch later entry being lost if mpv were killed (ouch).

This bug was discovered by looking at the code with a fresh pair of eyes
and thankfully not discovered through experience!